Use hidden friends properly
